refactor: parse args as paths, separate module for grid processing
